+++ phyloframe/legacy/_alifestd_unfurl_traversal_postorder_asexual.py
@@ -139,9 +139,9 @@ def alifestd_unfurl_traversal_postorder_asexual(
             and "next_sibling_id" in phylogeny_df.columns
         ):
             return _alifestd_unfurl_traversal_postorder_asexual_sibling_jit(
-                ancestor_ids.astype(np.int64),
-                phylogeny_df["first_child_id"].to_numpy().astype(np.int64),
-                phylogeny_df["next_sibling_id"].to_numpy().astype(np.int64),
+                ancestor_ids,
+                phylogeny_df["first_child_id"].to_numpy(),
+                phylogeny_df["next_sibling_id"].to_numpy(),
             )
         if (
             "csr_offsets" in phylogeny_df.columns
@@ -149,10 +149,10 @@ def alifestd_unfurl_traversal_postorder_asexual(
             and "num_children" in phylogeny_df.columns
         ):
             return _alifestd_unfurl_traversal_postorder_contiguous_asexual_jit(
-                ancestor_ids.astype(np.int64),
-                phylogeny_df["csr_offsets"].to_numpy().astype(np.int64),
-                phylogeny_df["csr_children"].to_numpy().astype(np.int64),
-                phylogeny_df["num_children"].to_numpy().astype(np.int64),
+                ancestor_ids,
+                phylogeny_df["csr_offsets"].to_numpy(),
+                phylogeny_df["csr_children"].to_numpy(),
+                phylogeny_df["num_children"].to_numpy(),
             )
         if "node_depth" not in phylogeny_df.columns:
             node_depths = _alifestd_calc_node_depth_asexual_contiguous(
